L-完善功能。代码编辑器已经可以从聊天气泡打开。

+++ CAC.client/Pages/CodeEditorPage/CodeEditSessionInfo.cs
@@ -12,7 +12,7 @@ namespace CAC.client.CodeEditorPage
     class CodeEditSessionInfo
     {
         public ContactItemViewModel Contact { get; set; }
-        public int MessageID { get; set; }
+        public long MessageID { get; set; }
         public string Code { get; set; }
         public string Language { get; set; }
         public DateTime CreateTime { get; set; }
